+++ src/themes/default/components/core/blocks/Microcart/Product.vue
@@ -9,14 +9,7 @@
       <div>
         <router-link
           class="serif h4 name"
-          :to="localizedRoute({
-            name: product.type_id + '-product',
-            params: {
-              parentSku: product.parentSku ? product.parentSku : product.sku,
-              slug: product.slug,
-              childSku: product.sku
-            }
-          })"
+          :to="productLink"
           data-testid="productLink"
           @click.native="$store.commit('ui/setMicrocart', false)"
         >
@@ -93,6 +86,8 @@
 <script>
 import config from 'config'
 import Product from '@vue-storefront/core/compatibility/components/blocks/Microcart/Product'
+import { currentStoreView } from '@vue-storefront/core/lib/multistore'
+import { formatProductLink } from '@vue-storefront/core/modules/url/helpers'
 
 import ProductImage from 'theme/components/core/ProductImage'
 import RemoveButton from './RemoveButton'
@@ -115,6 +110,9 @@ export default {
         loading: this.thumbnail,
         src: this.thumbnail
       }
+    },
+    productLink () {
+      return formatProductLink(this.product, currentStoreView().storeCode)
     }
   },
   data () {